+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added hermes handler to notify CRM about refreshing user's data. You can enable the feature in `config.local.neon` (see example file for reference). remp/web#2061
 - Fixed typo in the `package.json` definition for `moment-timezone` causing Yarn3 installation issues. [remp2020/mailer-module#2](https://github.com/remp2020/mailer-module/pull/2) 
 - Added new emit of `user-subscribed-variant` and `user-unsubscribed-variant` hermes event when user subscribe or unsubscribes from the mail type variant. remp/web#2061
+- Added `mail:unsubscribe-inactive-users` command to mailer module. remp/remp#1277
 
 ### [Sso]